requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
Pillow>=10.0.0
imagehash>=4.3.0
python-dateutil>=2.8.0
ExifRead>=3.0.0

# Optional for embeddings (uncomment if needed)
# transformers>=4.30.0
# torch>=2.0.0
# open-clip-torch>=2.20.0
# faiss-cpu>=1.7.0